refactor(animation): [iOS] Move FloatValueAnimator logic to shared base class
